% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Get parameters from a yaml file based on the current git branch}
\usage{
get_params(path)
}
\arguments{
\item{path}{String indicating path to the yaml file.}
}
\description{
Get parameters from a yaml file based on the current git branch
}
