% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_surveycto}
\alias{sync_surveycto}
\title{Sync data from SurveyCTO to a data warehouse}
\usage{
sync_surveycto(scto_params, wh_params)
}
\arguments{
\item{scto_params}{List of parameters for SurveyCTO.}

\item{wh_params}{List of parameters for the data warehouse.}
}
\description{
Sync data from SurveyCTO to a data warehouse
}
