% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Get parameters from a yaml file}
\usage{
get_params(path)
}
\arguments{
\item{path}{String indicating path to the yaml file.}
}
\description{
If GitHub Actions is running on the main branch, use prod. Otherwise dev.
}
